requirements.txt
alembic==1.4.2
anyio==3.5.0
APScheduler==3.6.3
backcall==0.2.0
cachetools==4.2.2
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.1
distro==1.7.0
entrypoints==0.3
et-xmlfile==1.1.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==2.8
ipykernel==6.6.1
ipython==7.31.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jupyter-client==7.1.0
jupyter-core==4.9.1
Mako==1.1.6
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numpy==1.22.0
onesignal-sdk==2.0.0
openpyxl==3.0.9
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.24
psycopg2==2.9.3
Pygments==2.11.2
python-dateutil==2.8.2
python-editor==1.0.4
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pyzmq==22.3.0
requests==2.22.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.3.11
tabula-py==2.3.0
tornado==6.1
traitlets==5.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.25.11
wcwidth==0.2.5
Werkzeug==2.0.2
XlsxWriter==3.0.3
